revise product application

+++ ControllerPage/Form1.cs
@@ -377,7 +377,8 @@ namespace ControllerPage
         private void Btn_Check_Click(object sender, EventArgs e)
         {
             Btn_Check.Enabled = false;
-            /*if (Button_Interface.Text == "RS-232")
+            
+            if (Button_Interface.Text == "RS-232")
             {
                 ProcessStartInfo procStartInfo1 = new ProcessStartInfo("/usr/bin/sudo", "systemctl stop modbusserver.service");
                 procStartInfo1.RedirectStandardOutput = true;
@@ -417,7 +418,7 @@ namespace ControllerPage
                 proc4.StartInfo = procStartInfo4;
                 proc4.Start();
             }
-            */
+            
             if (Button_Interface.Text != "RS-232" && Button_Interface.Text != "RS-485")
             {
                 MessageBox.Show("Please Pick Interface " + Button_Interface.Text, application_name);
@@ -1031,7 +1032,7 @@ namespace ControllerPage
                     {
                         combox_typemeasure = "Long_Sticky_Paddy";
                     }
-                    else if (ButtonProduct.Text == "Beras Long Parboiled" || ButtonProduct.Text == "ข้าวเปลือกนึ่ง")
+                    else if (ButtonProduct.Text == "Beras Long Parboiled" || ButtonProduct.Text == "ข้าวนึ่ง")
                     {
                         combox_typemeasure = "Long_Parboiled_Rice";
                     }
@@ -1140,7 +1141,7 @@ namespace ControllerPage
                     {
                         combox_typemeasure = "Long_Sticky_Paddy";
                     }
-                    else if (ButtonProduct.Text == "Beras Long Parboiled" || ButtonProduct.Text == "ข้าวเปลือกนึ่ง")
+                    else if (ButtonProduct.Text == "Beras Long Parboiled" || ButtonProduct.Text == "ข้าวนึ่ง")
                     {
                         combox_typemeasure = "Long_Parboiled_Rice";
                     }
@@ -1249,7 +1250,7 @@ namespace ControllerPage
                     {
                         combox_typemeasure = "Long_Sticky_Paddy";
                     }
-                    else if (ButtonProduct.Text == "Beras Long Parboiled" || ButtonProduct.Text == "ข้าวเปลือกนึ่ง")
+                    else if (ButtonProduct.Text == "Beras Long Parboiled" || ButtonProduct.Text == "ข้าวนึ่ง")
                     {
                         combox_typemeasure = "Long_Parboiled_Rice";
                     }
